+++ src/views/AuthoritiesSearch/AuthoritiesSearch.js
@@ -346,12 +346,22 @@ const AuthoritiesSearch = ({
         onClick={e => e.stopPropagation()}
       />
     ),
-    [searchResultListColumns.NUMBER_OF_TITLES]: ({ authRefType, numberOfTitles }) => {
+    /* eslint-disable-next-line */
+    [searchResultListColumns.NUMBER_OF_TITLES]: ({ id, authRefType, numberOfTitles }) => {
       if (authRefType !== AUTH_REF_TYPES.AUTHORIZED || numberOfTitles === 0) {
         return null;
       }
 
-      return numberOfTitles;
+      return (
+        <TextLink
+          to={`/inventory?qindex=authorityId&query=${id}&sort=title`}
+          target="_blank"
+          onClick={e => e.stopPropagation()}
+          data-testid="link-number-of-titles"
+        >
+          {numberOfTitles}
+        </TextLink>
+      );
     },
   };